+++ src/wreck/exceptions.py
@@ -4,8 +4,9 @@
 Package wide exceptions
 
 .. py:data:: __all__
-   :type: tuple[str, str]
-   :value: ("PyProjectTOMLParseError", "MissingRequirementsFoldersFiles")
+   :type: tuple[str, str, str, str]
+   :value: ("ArbitraryEqualityNotImplemented", "MissingPackageBaseFolder", \
+   "MissingRequirementsFoldersFiles", "PinMoreThanTwoSpecifiers")
 
    Module exports
 
@@ -13,8 +14,10 @@ Package wide exceptions
 
 __package__ = "wreck"
 __all__ = (
+    "ArbitraryEqualityNotImplemented",
     "MissingPackageBaseFolder",
     "MissingRequirementsFoldersFiles",
+    "PinMoreThanTwoSpecifiers",
 )
 
 
@@ -35,6 +38,36 @@ class MissingRequirementsFoldersFiles(AssertionError):
 
 class MissingPackageBaseFolder(AssertionError):
     """Loader did not provide package base folder. Do not know the cwd
+
+    :ivar msg: The error message
+    :vartype msg: str
+    """
+
+    def __init__(self, msg: str) -> None:
+        """Class constructor."""
+        super().__init__(msg)
+
+
+class ArbitraryEqualityNotImplemented(NotImplementedError):
+    """``===`` operator is not yet supported.
+
+    Convert this exception into an UnResolvable. End user
+    would have to manually handle the issue.
+
+    :ivar msg: The error message
+    :vartype msg: str
+    """
+
+    def __init__(self, msg: str) -> None:
+        """Class constructor."""
+        super().__init__(msg)
+
+
+class PinMoreThanTwoSpecifiers(NotImplementedError):
+    """A pin has ``>2`` specifiers. e.g. >=2.5.6, !=2.5.7, <2.7
+
+    Convert this exception into an UnResolvable. End user
+    would have to manually handle the issue.
 
     :ivar msg: The error message
     :vartype msg: str